+++ js/callback_feedly.js
@@ -1,5 +1,5 @@
 var reCode = /code=([^&]+)/;
 var reError = /error=([^&]+)/;
 
-opener.feedly.callback(location.href);
+parent.feedly.callback(location.href);
 window.close();