+++ Changelog.md
@@ -3,6 +3,7 @@
 - Font family for desktop/firefox mobile
 - Fix feedly published date
 - Fix feedly canonical href on some feeds
+- Use embedded iframe for feedly login (avoid use of window.open on Firefox Android and Desktop)
 
 
 # 0.2.5